Move the language selector into more generic location.
