+++ osu.Game.Rulesets.Karaoke/Screens/Edit/Components/UserInterface/LanguagesSelector.cs
@@ -18,11 +18,10 @@ using osu.Framework.Graphics.UserInterface;
 using osu.Game.Graphics.Sprites;
 using osu.Game.Overlays;
 using osu.Game.Rulesets.Karaoke.Graphics.UserInterfaceV2;
-using osu.Game.Rulesets.Karaoke.Screens.Edit.Components.UserInterface;
 using osu.Game.Rulesets.Karaoke.Utils;
 using osuTK;
 
-namespace osu.Game.Rulesets.Karaoke.Screens.Edit.Beatmaps.Configs.Generator.Components;
+namespace osu.Game.Rulesets.Karaoke.Screens.Edit.Components.UserInterface;
 
 public partial class LanguagesSelector : FillFlowContainer, IHasCurrentValue<CultureInfo[]>
 {
